pyproject.toml
[project]
name = "stockanalysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.1",
    "bcrypt==4.0.1",
    "fastapi==0.104.1",
    "httpx==0.25.2",
    "motor==3.3.1",
    "pandas==2.1.3",
    "passlib==1.7.4",
    "pydantic==2.5.2",
    "pymongo==4.6.0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.6",
    "uvicorn==0.24.0",
]
